refactor: split script into modules
