+++ app/templates/index.html
@@ -243,6 +243,6 @@
         </main>
     </div>
     
-    <script type="module" src="{{ url_for('static', filename='js/script.js') }}" defer></script>
+    <script type="module" src="{{ url_for('static', filename='js/main.js') }}" defer></script>
 </body>
 </html>
